Gather tests into one module
