+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -22,7 +22,7 @@ from polars.testing import (
 if TYPE_CHECKING:
     from zoneinfo import ZoneInfo
 
-    from polars.type_aliases import Ambiguous, PolarsTemporalType, StartBy, TimeUnit
+    from polars.type_aliases import Ambiguous, PolarsTemporalType, TimeUnit
 else:
     from polars.utils.convert import get_zoneinfo as ZoneInfo
 
@@ -551,90 +551,6 @@ def test_datetime_comp_tz_aware_invalid() -> None:
         _ = a > other
 
 
-@pytest.mark.parametrize("tzinfo", [None, ZoneInfo("Asia/Kathmandu")])
-def test_truncate_negative_offset(tzinfo: ZoneInfo | None) -> None:
-    time_zone = tzinfo.key if tzinfo is not None else None
-    df = pl.DataFrame(
-        {
-            "event_date": [
-                datetime(2021, 4, 11),
-                datetime(2021, 4, 29),
-                datetime(2021, 5, 29),
-            ],
-            "adm1_code": [1, 2, 1],
-        }
-    ).set_sorted("event_date")
-    df = df.with_columns(pl.col("event_date").dt.replace_time_zone(time_zone))
-    out = df.group_by_dynamic(
-        index_column="event_date",
-        every="1mo",
-        period="2mo",
-        offset="-1mo",
-        include_boundaries=True,
-    ).agg(
-        [
-            pl.col("adm1_code"),
-        ]
-    )
-
-    assert out["event_date"].to_list() == [
-        datetime(2021, 3, 1, tzinfo=tzinfo),
-        datetime(2021, 4, 1, tzinfo=tzinfo),
-        datetime(2021, 5, 1, tzinfo=tzinfo),
-    ]
-    df = pl.DataFrame(
-        {
-            "event_date": [
-                datetime(2021, 4, 11),
-                datetime(2021, 4, 29),
-                datetime(2021, 5, 29),
-            ],
-            "adm1_code": [1, 2, 1],
-            "five_type": ["a", "b", "a"],
-            "actor": ["a", "a", "a"],
-            "admin": ["a", "a", "a"],
-            "fatalities": [10, 20, 30],
-        }
-    ).set_sorted("event_date")
-    df = df.with_columns(pl.col("event_date").dt.replace_time_zone(time_zone))
-
-    out = df.group_by_dynamic(
-        index_column="event_date",
-        every="1mo",
-        by=["admin", "five_type", "actor"],
-    ).agg([pl.col("adm1_code").unique(), (pl.col("fatalities") > 0).sum()])
-
-    assert out["event_date"].to_list() == [
-        datetime(2021, 4, 1, tzinfo=tzinfo),
-        datetime(2021, 5, 1, tzinfo=tzinfo),
-        datetime(2021, 4, 1, tzinfo=tzinfo),
-    ]
-
-    for dt in [pl.Int32, pl.Int64]:
-        df = (
-            pl.DataFrame(
-                {
-                    "idx": np.arange(6),
-                    "A": ["A", "A", "B", "B", "B", "C"],
-                }
-            )
-            .with_columns(pl.col("idx").cast(dt))
-            .set_sorted("idx")
-        )
-
-        out = df.group_by_dynamic(
-            "idx", every="2i", period="3i", include_boundaries=True
-        ).agg(pl.col("A"))
-
-        assert out.shape == (4, 4)
-        assert out["A"].to_list() == [
-            ["A"],
-            ["A", "A", "B"],
-            ["B", "B", "B"],
-            ["B", "C"],
-        ]
-
-
 def test_to_arrow() -> None:
     date_series = pl.Series("dates", ["2022-01-16", "2022-01-17"]).str.strptime(
         pl.Date, "%Y-%m-%d"
@@ -676,84 +592,6 @@ def test_explode_date() -> None:
             ("b", dclass(2021, 12, 1), None),
             ("b", dclass(2021, 12, 1), 0.25),
         ]
-
-
-def test_groupy_by_dynamic_median_10695() -> None:
-    df = pl.DataFrame(
-        {
-            "timestamp": pl.datetime_range(
-                datetime(2023, 8, 22, 15, 44, 30),
-                datetime(2023, 8, 22, 15, 48, 50),
-                "20s",
-                eager=True,
-            ),
-            "foo": [2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-        }
-    )
-
-    assert df.group_by_dynamic(
-        index_column="timestamp",
-        every="60s",
-        period="3m",
-    ).agg(
-        pl.col("foo").median()
-    ).to_dict(False) == {
-        "timestamp": [
-            datetime(2023, 8, 22, 15, 43),
-            datetime(2023, 8, 22, 15, 44),
-            datetime(2023, 8, 22, 15, 45),
-            datetime(2023, 8, 22, 15, 46),
-            datetime(2023, 8, 22, 15, 47),
-            datetime(2023, 8, 22, 15, 48),
-        ],
-        "foo": [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
-    }
-
-
-def test_group_by_dynamic_when_conversion_crosses_dates_7274() -> None:
-    df = (
-        pl.DataFrame(
-            data={
-                "timestamp": ["1970-01-01 00:00:00+01:00", "1970-01-01 01:00:00+01:00"],
-                "value": [1, 1],
-            }
-        )
-        .with_columns(
-            pl.col("timestamp")
-            .str.strptime(pl.Datetime, format="%Y-%m-%d %H:%M:%S%:z")
-            .dt.convert_time_zone("Africa/Lagos")
-            .set_sorted()
-        )
-        .with_columns(
-            pl.col("timestamp")
-            .dt.convert_time_zone("UTC")
-            .alias("timestamp_utc")
-            .set_sorted()
-        )
-    )
-    result = df.group_by_dynamic(
-        index_column="timestamp", every="1d", closed="left"
-    ).agg(pl.col("value").count())
-    expected = pl.DataFrame({"timestamp": [datetime(1970, 1, 1)], "value": [2]})
-    expected = expected.with_columns(
-        pl.col("timestamp").dt.replace_time_zone("Africa/Lagos"),
-        pl.col("value").cast(pl.UInt32),
-    )
-    assert_frame_equal(result, expected)
-    result = df.group_by_dynamic(
-        index_column="timestamp_utc", every="1d", closed="left"
-    ).agg(pl.col("value").count())
-    expected = pl.DataFrame(
-        {
-            "timestamp_utc": [datetime(1969, 12, 31), datetime(1970, 1, 1)],
-            "value": [1, 1],
-        }
-    )
-    expected = expected.with_columns(
-        pl.col("timestamp_utc").dt.replace_time_zone("UTC"),
-        pl.col("value").cast(pl.UInt32),
-    )
-    assert_frame_equal(result, expected)
 
 
 def test_rolling() -> None:
@@ -952,185 +790,6 @@ def test_read_utc_times_parquet() -> None:
     df_in = pl.read_parquet(f)
     tz = ZoneInfo("UTC")
     assert df_in["Timestamp"][0] == datetime(2022, 1, 1, 0, 0, tzinfo=tz)
-
-
-@pytest.mark.parametrize("time_zone", [None, "Asia/Kathmandu"])
-def test_default_negative_every_offset_dynamic_group_by(time_zone: str | None) -> None:
-    # 2791
-    dts = [
-        datetime(2020, 1, 1),
-        datetime(2020, 1, 2),
-        datetime(2020, 2, 1),
-        datetime(2020, 3, 1),
-    ]
-    df = pl.DataFrame({"dt": dts, "idx": range(len(dts))}).set_sorted("dt")
-    df = df.with_columns(pl.col("dt").dt.replace_time_zone(time_zone))
-    out = df.group_by_dynamic(index_column="dt", every="1mo", closed="right").agg(
-        pl.col("idx")
-    )
-
-    expected = pl.DataFrame(
-        {
-            "dt": [
-                datetime(2019, 12, 1, 0, 0),
-                datetime(2020, 1, 1, 0, 0),
-                datetime(2020, 2, 1, 0, 0),
-            ],
-            "idx": [[0], [1, 2], [3]],
-        }
-    )
-    expected = expected.with_columns(pl.col("dt").dt.replace_time_zone(time_zone))
-    assert_frame_equal(out, expected)
-
-
-@pytest.mark.parametrize(
-    ("rule", "offset"),
-    [
-        ("1h", timedelta(hours=2)),
-        ("1d", timedelta(days=2)),
-        ("1w", timedelta(weeks=2)),
-    ],
-)
-def test_group_by_dynamic_crossing_dst(rule: str, offset: timedelta) -> None:
-    start_dt = datetime(2021, 11, 7)
-    end_dt = start_dt + offset
-    date_range = pl.datetime_range(
-        start_dt, end_dt, rule, time_zone="US/Central", eager=True
-    )
-    df = pl.DataFrame({"time": date_range, "value": range(len(date_range))})
-    result = df.group_by_dynamic("time", every=rule, start_by="datapoint").agg(
-        pl.col("value").mean()
-    )
-    expected = pl.DataFrame(
-        {"time": date_range, "value": range(len(date_range))},
-        schema_overrides={"value": pl.Float64},
-    )
-    assert_frame_equal(result, expected)
-
-
-@pytest.mark.parametrize(
-    ("start_by", "expected_time", "expected_value"),
-    [
-        (
-            "monday",
-            [
-                datetime(2021, 11, 1),
-                datetime(2021, 11, 8),
-            ],
-            [0.0, 4.0],
-        ),
-        (
-            "tuesday",
-            [
-                datetime(2021, 11, 2),
-                datetime(2021, 11, 9),
-            ],
-            [0.5, 4.5],
-        ),
-        (
-            "wednesday",
-            [
-                datetime(2021, 11, 3),
-                datetime(2021, 11, 10),
-            ],
-            [1.0, 5.0],
-        ),
-        (
-            "thursday",
-            [
-                datetime(2021, 11, 4),
-                datetime(2021, 11, 11),
-            ],
-            [1.5, 5.5],
-        ),
-        (
-            "friday",
-            [
-                datetime(2021, 11, 5),
-                datetime(2021, 11, 12),
-            ],
-            [2.0, 6.0],
-        ),
-        (
-            "saturday",
-            [
-                datetime(2021, 11, 6),
-                datetime(2021, 11, 13),
-            ],
-            [2.5, 6.5],
-        ),
-        (
-            "sunday",
-            [
-                datetime(2021, 11, 7),
-                datetime(2021, 11, 14),
-            ],
-            [3.0, 7.0],
-        ),
-    ],
-)
-def test_group_by_dynamic_startby_monday_crossing_dst(
-    start_by: StartBy, expected_time: list[datetime], expected_value: list[float]
-) -> None:
-    start_dt = datetime(2021, 11, 7)
-    end_dt = datetime(2021, 11, 14)
-    date_range = pl.datetime_range(
-        start_dt, end_dt, "1d", time_zone="US/Central", eager=True
-    )
-    df = pl.DataFrame({"time": date_range, "value": range(len(date_range))})
-    result = df.group_by_dynamic("time", every="1w", start_by=start_by).agg(
-        pl.col("value").mean()
-    )
-    expected = pl.DataFrame(
-        {"time": expected_time, "value": expected_value},
-    )
-    expected = expected.with_columns(pl.col("time").dt.replace_time_zone("US/Central"))
-    assert_frame_equal(result, expected)
-
-
-def test_group_by_dynamic_startby_monday_dst_8737() -> None:
-    start_dt = datetime(2021, 11, 6, 20)
-    stop_dt = datetime(2021, 11, 7, 20)
-    date_range = pl.datetime_range(
-        start_dt, stop_dt, "1d", time_zone="US/Central", eager=True
-    )
-    df = pl.DataFrame({"time": date_range, "value": range(len(date_range))})
-    result = df.group_by_dynamic("time", every="1w", start_by="monday").agg(
-        pl.col("value").mean()
-    )
-    expected = pl.DataFrame(
-        {
-            "time": [
-                datetime(2021, 11, 1),
-            ],
-            "value": [0.5],
-        },
-    )
-    expected = expected.with_columns(pl.col("time").dt.replace_time_zone("US/Central"))
-    assert_frame_equal(result, expected)
-
-
-def test_group_by_dynamic_monthly_crossing_dst() -> None:
-    start_dt = datetime(2021, 11, 1)
-    end_dt = datetime(2021, 12, 1)
-    date_range = pl.datetime_range(
-        start_dt, end_dt, "1mo", time_zone="US/Central", eager=True
-    )
-    df = pl.DataFrame({"time": date_range, "value": range(len(date_range))})
-    result = df.group_by_dynamic("time", every="1mo").agg(pl.col("value").mean())
-    expected = pl.DataFrame(
-        {"time": date_range, "value": range(len(date_range))},
-        schema_overrides={"value": pl.Float64},
-    )
-    assert_frame_equal(result, expected)
-
-
-def test_group_by_dynamic_2d_9333() -> None:
-    df = pl.DataFrame({"ts": [datetime(2000, 1, 1, 3)], "values": [10.0]})
-    df = df.with_columns(pl.col("ts").set_sorted())
-    result = df.group_by_dynamic("ts", every="2d").agg(pl.col("values"))
-    expected = pl.DataFrame({"ts": [datetime(1999, 12, 31, 0)], "values": [[10.0]]})
-    assert_frame_equal(result, expected)
 
 
 def test_asof_join_tolerance_grouper() -> None:
@@ -3200,48 +2859,6 @@ def test_misc_precision_any_value_conversion(time_zone: Any, warn: bool) -> None
 def test_pytime_conversion(tm: time) -> None:
     s = pl.Series("tm", [tm])
     assert s.to_list() == [tm]
-
-
-@pytest.mark.parametrize(
-    ("input_df", "expected_grouped_df"),
-    [
-        (
-            (
-                pl.DataFrame(
-                    {
-                        "dt": [
-                            datetime(2021, 12, 31, 0, 0, 0),
-                            datetime(2022, 1, 1, 0, 0, 1),
-                            datetime(2022, 3, 31, 0, 0, 1),
-                            datetime(2022, 4, 1, 0, 0, 1),
-                        ]
-                    }
-                )
-            ),
-            pl.DataFrame(
-                {
-                    "dt": [
-                        datetime(2021, 10, 1),
-                        datetime(2022, 1, 1),
-                        datetime(2022, 4, 1),
-                    ],
-                    "num_points": [1, 2, 1],
-                },
-                schema={"dt": pl.Datetime, "num_points": pl.UInt32},
-            ).sort("dt"),
-        )
-    ],
-)
-def test_group_by_dynamic(
-    input_df: pl.DataFrame, expected_grouped_df: pl.DataFrame
-) -> None:
-    result = (
-        input_df.sort("dt")
-        .group_by_dynamic("dt", every="1q")
-        .agg(pl.col("dt").count().alias("num_points"))
-        .sort("dt")
-    )
-    assert_frame_equal(result, expected_grouped_df)
 
 
 def test_group_by_rolling_duplicates() -> None:
