+++ py-polars/tests/unit/operations/rolling/test_rolling.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import sys
 from datetime import date, datetime, timedelta
 from typing import TYPE_CHECKING
 
@@ -8,20 +7,12 @@ import numpy as np
 import pytest
 from numpy import nan
 
-from polars.exceptions import ComputeError
-
-if sys.version_info >= (3, 9):
-    from zoneinfo import ZoneInfo
-else:
-    # Import from submodule due to typing issue with backports.zoneinfo package:
-    # https://github.com/pganssle/zoneinfo/issues/125
-    from backports.zoneinfo._zoneinfo import ZoneInfo
-
 import polars as pl
+from polars.exceptions import ComputeError
 from polars.testing import assert_frame_equal, assert_series_equal
 
 if TYPE_CHECKING:
-    from polars.type_aliases import ClosedInterval, Label, TimeUnit
+    from polars.type_aliases import ClosedInterval, TimeUnit
 
 
 @pytest.fixture()
@@ -408,30 +399,6 @@ def test_rolling_slice_pushdown() -> None:
     }
 
 
-def test_group_by_dynamic_slice_pushdown() -> None:
-    df = pl.DataFrame({"a": [1, 2, 3], "b": ["a", "a", "b"], "c": [1, 3, 5]}).lazy()
-    df = (
-        df.sort("a")
-        .group_by_dynamic(
-            "a",
-            by="b",
-            every="2i",
-        )
-        .agg(
-            [
-                (pl.col("c") - pl.col("c").shift_and_fill(fill_value=0, periods=1))
-                .sum()
-                .alias("c")
-            ]
-        )
-    )
-    assert df.head(2).collect().to_dict(False) == {
-        "b": ["a", "a"],
-        "a": [0, 2],
-        "c": [1, 3],
-    }
-
-
 def test_overlapping_groups_4628() -> None:
     df = pl.DataFrame(
         {
@@ -506,257 +473,6 @@ def test_rolling_var_numerical_stability_5197() -> None:
     assert res[:4] == [None] * 4
 
 
-@pytest.mark.parametrize(
-    ("every", "offset"),
-    [
-        ("3d", "-1d"),
-        (timedelta(days=3), timedelta(days=-1)),
-    ],
-)
-def test_dynamic_group_by_timezone_awareness(
-    every: str | timedelta, offset: str | timedelta
-) -> None:
-    df = pl.DataFrame(
-        (
-            pl.datetime_range(
-                datetime(2020, 1, 1),
-                datetime(2020, 1, 10),
-                timedelta(days=1),
-                time_unit="ns",
-                eager=True,
-            )
-            .alias("datetime")
-            .dt.replace_time_zone("UTC"),
-            pl.arange(1, 11, eager=True).alias("value"),
-        )
-    )
-
-    assert (
-        df.group_by_dynamic(
-            "datetime",
-            every=every,
-            offset=offset,
-            closed="right",
-            include_boundaries=True,
-            label="datapoint",
-        ).agg(pl.col("value").last())
-    ).dtypes == [pl.Datetime("ns", "UTC")] * 3 + [pl.Int64]
-
-
-@pytest.mark.parametrize("tzinfo", [None, ZoneInfo("Asia/Kathmandu")])
-def test_group_by_dynamic_startby_5599(tzinfo: ZoneInfo | None) -> None:
-    # start by datapoint
-    start = datetime(2022, 12, 16, tzinfo=tzinfo)
-    stop = datetime(2022, 12, 16, hour=3, tzinfo=tzinfo)
-    df = pl.DataFrame({"date": pl.datetime_range(start, stop, "30m", eager=True)})
-
-    assert df.group_by_dynamic(
-        "date",
-        every="31m",
-        include_boundaries=True,
-        label="datapoint",
-        start_by="datapoint",
-    ).agg(pl.count()).to_dict(False) == {
-        "_lower_boundary": [
-            datetime(2022, 12, 16, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 0, 31, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 2, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 33, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 4, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 35, tzinfo=tzinfo),
-        ],
-        "_upper_boundary": [
-            datetime(2022, 12, 16, 0, 31, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 2, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 33, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 4, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 35, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 3, 6, tzinfo=tzinfo),
-        ],
-        "date": [
-            datetime(2022, 12, 16, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 0, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 1, 30, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 0, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 2, 30, tzinfo=tzinfo),
-            datetime(2022, 12, 16, 3, 0, tzinfo=tzinfo),
-        ],
-        "count": [2, 1, 1, 1, 1, 1],
-    }
-
-    # start by monday
-    start = datetime(2022, 1, 1, tzinfo=tzinfo)
-    stop = datetime(2022, 1, 12, 7, tzinfo=tzinfo)
-
-    df = pl.DataFrame(
-        {"date": pl.datetime_range(start, stop, "12h", eager=True)}
-    ).with_columns(pl.col("date").dt.weekday().alias("day"))
-
-    result = df.group_by_dynamic(
-        "date",
-        every="1w",
-        period="3d",
-        include_boundaries=True,
-        start_by="monday",
-        label="datapoint",
-    ).agg([pl.count(), pl.col("day").first().alias("data_day")])
-    assert result.to_dict(False) == {
-        "_lower_boundary": [
-            datetime(2022, 1, 3, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 10, 0, 0, tzinfo=tzinfo),
-        ],
-        "_upper_boundary": [
-            datetime(2022, 1, 6, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 13, 0, 0, tzinfo=tzinfo),
-        ],
-        "date": [
-            datetime(2022, 1, 3, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 10, 0, 0, tzinfo=tzinfo),
-        ],
-        "count": [6, 5],
-        "data_day": [1, 1],
-    }
-    # start by saturday
-    result = df.group_by_dynamic(
-        "date",
-        every="1w",
-        period="3d",
-        include_boundaries=True,
-        start_by="saturday",
-        label="datapoint",
-    ).agg([pl.count(), pl.col("day").first().alias("data_day")])
-    assert result.to_dict(False) == {
-        "_lower_boundary": [
-            datetime(2022, 1, 1, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 8, 0, 0, tzinfo=tzinfo),
-        ],
-        "_upper_boundary": [
-            datetime(2022, 1, 4, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 11, 0, 0, tzinfo=tzinfo),
-        ],
-        "date": [
-            datetime(2022, 1, 1, 0, 0, tzinfo=tzinfo),
-            datetime(2022, 1, 8, 0, 0, tzinfo=tzinfo),
-        ],
-        "count": [6, 6],
-        "data_day": [6, 6],
-    }
-
-
-def test_group_by_dynamic_by_monday_and_offset_5444() -> None:
-    df = pl.DataFrame(
-        {
-            "date": [
-                "2022-11-01",
-                "2022-11-02",
-                "2022-11-05",
-                "2022-11-08",
-                "2022-11-08",
-                "2022-11-09",
-                "2022-11-10",
-            ],
-            "label": ["a", "b", "a", "a", "b", "a", "b"],
-            "value": [1, 2, 3, 4, 5, 6, 7],
-        }
-    ).with_columns(pl.col("date").str.strptime(pl.Date, "%Y-%m-%d").set_sorted())
-
-    result = df.group_by_dynamic(
-        "date", every="1w", offset="1d", by="label", start_by="monday"
-    ).agg(pl.col("value").sum())
-
-    assert result.to_dict(False) == {
-        "label": ["a", "a", "b", "b"],
-        "date": [
-            date(2022, 11, 1),
-            date(2022, 11, 8),
-            date(2022, 11, 1),
-            date(2022, 11, 8),
-        ],
-        "value": [4, 10, 2, 12],
-    }
-
-    # test empty
-    result_empty = (
-        df.filter(pl.col("date") == date(1, 1, 1))
-        .group_by_dynamic(
-            "date", every="1w", offset="1d", by="label", start_by="monday"
-        )
-        .agg(pl.col("value").sum())
-    )
-    assert result_empty.schema == result.schema
-
-
-def test_group_by_dynamic_truncate_to_label_deprecation() -> None:
-    df = pl.LazyFrame({"ts": [], "n": []})
-    with pytest.warns(
-        DeprecationWarning, match="replace `truncate=False` with `label='datapoint'`"
-    ):
-        df.group_by_dynamic("ts", every="1d", truncate=False)
-    with pytest.warns(
-        DeprecationWarning, match="replace `truncate=True` with `label='left'`"
-    ):
-        df.group_by_dynamic("ts", every="1d", truncate=True)
-
-
-@pytest.mark.parametrize(
-    ("label", "expected"),
-    [
-        ("left", [datetime(2020, 1, 1), datetime(2020, 1, 2)]),
-        ("right", [datetime(2020, 1, 2), datetime(2020, 1, 3)]),
-        ("datapoint", [datetime(2020, 1, 1, 1), datetime(2020, 1, 2, 3)]),
-    ],
-)
-def test_group_by_dynamic_label(label: Label, expected: list[datetime]) -> None:
-    df = pl.DataFrame(
-        {
-            "ts": [
-                datetime(2020, 1, 1, 1),
-                datetime(2020, 1, 1, 2),
-                datetime(2020, 1, 2, 3),
-                datetime(2020, 1, 2, 4),
-            ],
-            "n": [1, 2, 3, 4],
-            "group": ["a", "a", "b", "b"],
-        }
-    ).sort("ts")
-    result = (
-        df.group_by_dynamic("ts", every="1d", label=label, by="group")
-        .agg(pl.col("n"))["ts"]
-        .to_list()
-    )
-    assert result == expected
-
-
-@pytest.mark.parametrize(
-    ("label", "expected"),
-    [
-        ("left", [datetime(2020, 1, 1), datetime(2020, 1, 2), datetime(2020, 1, 3)]),
-        ("right", [datetime(2020, 1, 2), datetime(2020, 1, 3), datetime(2020, 1, 4)]),
-        (
-            "datapoint",
-            [datetime(2020, 1, 1, 1), datetime(2020, 1, 2, 2), datetime(2020, 1, 3, 3)],
-        ),
-    ],
-)
-def test_group_by_dynamic_label_with_by(label: Label, expected: list[datetime]) -> None:
-    df = pl.DataFrame(
-        {
-            "ts": [
-                datetime(2020, 1, 1, 1),
-                datetime(2020, 1, 2, 2),
-                datetime(2020, 1, 3, 3),
-            ],
-            "n": [1, 2, 3],
-        }
-    ).sort("ts")
-    result = (
-        df.group_by_dynamic("ts", every="1d", label=label)
-        .agg(pl.col("n"))["ts"]
-        .to_list()
-    )
-    assert result == expected
-
-
 def test_group_by_rolling_iter() -> None:
     df = pl.DataFrame(
         {
@@ -819,25 +535,6 @@ def test_rolling_skew_window_offset() -> None:
     assert (pl.arange(0, 20, eager=True) ** 2).rolling_skew(20)[
         -1
     ] == 0.6612545648596286
-
-
-def test_rolling_kernels_group_by_dynamic_7548() -> None:
-    assert pl.DataFrame(
-        {"time": pl.arange(0, 4, eager=True), "value": pl.arange(0, 4, eager=True)}
-    ).group_by_dynamic("time", every="1i", period="3i").agg(
-        pl.col("value"),
-        pl.col("value").min().alias("min_value"),
-        pl.col("value").max().alias("max_value"),
-        pl.col("value").sum().alias("sum_value"),
-    ).to_dict(
-        False
-    ) == {
-        "time": [-1, 0, 1, 2, 3],
-        "value": [[0, 1], [0, 1, 2], [1, 2, 3], [2, 3], [3]],
-        "min_value": [0, 0, 1, 2, 3],
-        "max_value": [1, 2, 3, 3, 3],
-        "sum_value": [1, 3, 6, 5, 3],
-    }
 
 
 def test_rolling_cov_corr() -> None:
